added module, edit migration, etc

+++ database/migrations/2015_08_22_045657_create_purchasing_details_table.php
@@ -18,10 +18,10 @@ class CreatePurchasingDetailsTable extends Migration
             $table->integer('product_id')->unsigned();
             $table->string('product_name', 50);
             $table->string('unit', 10);
-            $table->integer('purchase_price')->unsigned();
-            $table->integer('discount')->unsigned();
-            $table->integer('quantity')->unsigned();
-            $table->integer('amount')->unsigned();
+            $table->decimal('purchase_price', 12, 2)->unsigned();
+            $table->decimal('discount', 5, 2)->unsigned();
+            $table->decimal('quantity', 10, 3)->unsigned();
+            $table->decimal('amount', 12, 2)->unsigned();
             $table->timestamps();
             $table->integer('created_by')->unsigned();
             $table->integer('updated_by')->unsigned();
